refactor(cc-addon-header): refactor api calls and create a `cc-addon-header.api.js` file
